feat(Toggle): rename SwitchField → ToggleField, ToggleField → ButtonToggle
- Rename SwitchField to ToggleField to align with SAIL's a!toggleField()
- Rename ToggleField (button-style toggle) to ButtonToggle
- Delete src/components/Switch/ directory
- Create src/components/ButtonToggle/ with renamed component
- Update src/components/Toggle/ to contain the switch-based implementation
- Update ApplicationHeader imports and usages
- Update src/components/index.ts exports
- Update AGENTS.md with new component quick reference patterns
- Rename switchLabelPosition prop to toggleLabelPosition

Closes #99
