+++ src/components/ButtonToggle/index.ts
@@ -1,0 +1,2 @@
+export { ButtonToggle } from './ButtonToggle'
+export type { ButtonToggleProps } from './ButtonToggle'
